- рефлизованны ставки по цветам с сохранением размера банка

+++ app/src/main/java/com/example/rulegamecompose/data/impl/RuleRepositoryImpl.kt
@@ -4,9 +4,9 @@ import com.example.rulegamecompose.data.storage.RuleStorage
 import com.example.rulegamecompose.domain.api.RuleRepositiry
 
 class RuleRepositoryImpl(private val storage: RuleStorage): RuleRepositiry {
-    override fun saveCash(value: Int) {
+    override fun saveCash(value: Float) {
         storage.setCash(value)
     }
 
-    override fun getCashValue(): Int = storage.getCash()
+    override fun getCashValue(): Float = storage.getCash()
 }
