+++ app/src/main/java/com/example/rulegamecompose/data/storage/RuleStorage.kt
@@ -1,7 +1,7 @@
 package com.example.rulegamecompose.data.storage
 
 interface RuleStorage {
-    fun getCash(): Int
+    fun getCash(): Float
 
-    fun setCash(value: Int)
+    fun setCash(value: Float)
 }
