+++ app/src/main/java/com/example/rulegamecompose/domain/api/RuleRepositiry.kt
@@ -1,6 +1,6 @@
 package com.example.rulegamecompose.domain.api
 
 interface RuleRepositiry {
-    fun saveCash(value: Int)
-    fun getCashValue(): Int
+    fun saveCash(value: Float)
+    fun getCashValue(): Float
 }
